refact: separating menu and project context
