+++ src/hooks/useProjectMenu.tsx
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { ProjectMenuContext } from 'src/context/ProjectMenuContext'
+
+export const useProjectMenu = () => useContext(ProjectMenuContext)
