+++ src/views/pages/projetos/configurar-projeto/Menu/Environments/AddInputDevice.tsx
@@ -16,6 +16,9 @@ import {
 import CustomTextField from 'src/@core/components/mui/text-field'
 
 import useGetDataApi from 'src/hooks/useGetDataApi'
+import { useProjectMenu } from 'src/hooks/useProjectMenu'
+
+import { checkPortName, checkSequenceIndex } from 'src/utils/project'
 
 import toast from 'react-hot-toast'
 
@@ -24,9 +27,6 @@ import { useForm, Controller } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 
 import { api } from 'src/services/api'
-
-import { useProject } from 'src/hooks/useProject'
-import { checkPortName, checkSequenceIndex } from 'src/utils/project'
 
 const schema = yup.object().shape({
   name: yup.string().required('Nome obrigatório'),
@@ -70,7 +70,7 @@ const AddInputDevice = ({
 
   const { id } = router.query
 
-  const { handleAvaliableInputPorts } = useProject()
+  const { handleAvaliableInputPorts } = useProjectMenu()
 
   const [ports, setPorts] = useState<any[]>([])
